# Get rid of Haml::Buffer