--- a/lib/haml/temple_engine.rb
+++ b/lib/haml/temple_engine.rb
@@ -2,11 +2,11 @@
 
 require 'temple'
 require 'haml/escapable'
-require 'haml/generator'
 
 module Haml
   class TempleEngine < Temple::Engine
     define_options(
+      :buffer_class,
       attr_wrapper:         "'",
       autoclose:            %w(area base basefont br col command embed frame
                                hr img input isindex keygen link menuitem meta
@@ -17,12 +17,14 @@ module Haml
       escape_filter_interpolations: nil,
       filename:             '(haml)',
       format:               :html5,
+      generator:            Temple::Generators::ArrayBuffer,
       hyphenate_data_attrs: true,
       line:                 1,
       mime_type:            'text/html',
       preserve:             %w(textarea pre code),
       remove_whitespace:    false,
       suppress_eval:        false,
+      use_html_safe:        false,
       cdata:                false,
       parser_class:         ::Haml::Parser,
       compiler_class:       ::Haml::Compiler,
@@ -36,7 +38,7 @@ module Haml
     filter :ControlFlow
     filter :MultiFlattener
     filter :StaticMerger
-    use Generator
+    use :Generator, -> { options[:generator] }
 
     def compile(template)
       initialize_encoding(template, options[:encoding])
@@ -55,31 +57,14 @@ module Haml
       return @precompiled.encode(encoding)
     end
 
-    def precompiled_with_return_value
-      "#{precompiled};#{precompiled_method_return_value}".dup
-    end
-
     # The source code that is evaluated to produce the Haml document.
     #
     # This is automatically converted to the correct encoding
     # (see {file:REFERENCE.md#encodings the `:encoding` option}).
     #
     # @return [String]
-    def precompiled_with_ambles(local_names, after_preamble: '')
-      preamble = <<END.tr("\n", ';')
-begin
-extend Haml::Helpers
-_hamlout = @haml_buffer = Haml::Buffer.new(haml_buffer, #{Options.new(options).for_buffer.inspect})
-_erbout = _hamlout.buffer
-#{after_preamble}
-END
-      postamble = <<END.tr("\n", ';')
-#{precompiled_method_return_value}
-ensure
-@haml_buffer = @haml_buffer.upper if @haml_buffer
-end
-END
-      "#{preamble}#{locals_code(local_names)}#{precompiled}#{postamble}".dup
+    def precompiled_with_ambles(local_names = [])
+      "#{locals_code(local_names)}#{precompiled}".dup
     end
 
     private
@@ -90,12 +75,6 @@ END
       else
         @encoding = Encoding.default_internal || template.encoding
       end
-    end
-
-    # Returns the string used as the return value of the precompiled method.
-    # This method exists so it can be monkeypatched to return modified values.
-    def precompiled_method_return_value
-      "_erbout"
     end
 
     def locals_code(names)