--- a/test/results/render_layout.xhtml
+++ b/test/results/render_layout.xhtml
@@ -1,3 +1,0 @@
-Before
-During
-After