--- a/test/templates/_partial.haml
+++ b/test/templates/_partial.haml
@@ -1,8 +1,0 @@
-%p
-  @foo = 
-  = @foo
-- @foo = 'value three'
-== Toplevel? #{haml_buffer.toplevel?}
-%p
-  @foo = 
-  = @foo