--- a/test/templates/partial_layout_erb.erb
+++ b/test/templates/partial_layout_erb.erb
@@ -1,4 +1,0 @@
-<h1>Partial layout used with for block:</h1>
-<%= render :layout => 'layout_for_partial' do -%>
-Some content within a layout
-<% end %>