--- a/test/templates/render_layout.haml
+++ b/test/templates/render_layout.haml
@@ -1,2 +1,0 @@
-= render :layout => 'layout' do
-  During