Same context as https://github.com/haml/haml/pull/1057#issue-663055925. This breaks a lot of things:

* Most of Rails monkey patches
  * Including HTML escape methods. Ones supporting Rails are named `_safe`.
* `haml_buffer`, `_hamlout`, and `_erbout`
* `encoding` option of `Haml::Engine`

I also haven't tested this well. We probably need to test and fix this a lot, and the release of this must start from beta and it should be left in beta for a while.

## Benchmark
I used [a benchmark based on Slim's](https://github.com/k0kubun/hamlit/blob/master/benchmark/slim/run-benchmarks.rb) with Ruby 2.7.3, and this PR gives the following difference. For optimal performance, we also need ["Limited attributes hyphenation" and "Limited boolean attributes"](https://github.com/k0kubun/hamlit/blob/master/REFERENCE.md#limitations), but I think that should be done in another PR.

### Before
```
-------------------------------------------------
       erubi v1.10.0    323.229k (± 0.4%) i/s -      1.632M
         slim v4.1.0    322.791k (± 0.1%) i/s -      1.633M
         haml v5.2.1    190.768k (± 1.5%) i/s -    955.098k
         faml v0.8.1    276.984k (± 0.2%) i/s -      1.404M
      hamlit v2.15.0    347.165k (± 0.3%) i/s -      1.759M

Comparison:
      hamlit v2.15.0:   347164.6 i/s
       erubi v1.10.0:   323229.5 i/s - 1.07x slower
         slim v4.1.0:   322791.3 i/s - 1.08x slower
         faml v0.8.1:   276984.2 i/s - 1.25x slower
         haml v5.2.1:   190767.5 i/s - 1.82x slower
```

### After
```
-------------------------------------------------
       erubi v1.10.0    347.070k (± 0.8%) i/s -      1.762M
         slim v4.1.0    337.808k (± 0.5%) i/s -      1.718M
         haml v5.2.1    246.472k (± 0.4%) i/s -      1.240M
         faml v0.8.1    288.625k (± 0.4%) i/s -      1.461M
      hamlit v2.15.0    367.851k (± 0.3%) i/s -      1.846M

Comparison:
      hamlit v2.15.0:   367850.7 i/s
       erubi v1.10.0:   347069.9 i/s - 1.06x slower
         slim v4.1.0:   337807.6 i/s - 1.09x slower
         faml v0.8.1:   288625.1 i/s - 1.27x slower
         haml v5.2.1:   246472.2 i/s - 1.49x slower
```